+++ R/get_statistic.R
@@ -345,14 +345,16 @@ get_statistic.vglm <- function(x, ...) {
 get_statistic.vgam <- function(x, ...) {
   params <- get_parameters(x)
   out <- data.frame(
-    Parameter = params$Parameter,
-    Statistic = as.vector(params$Estimate / sqrt(diag(get_varcov(x)))),
-    Component = params$Component,
+    Parameter = names(x@nl.chisq),
+    Statistic = x@nl.chisq,
     stringsAsFactors = FALSE,
     row.names = NULL
   )
 
-  out <- .remove_backticks_from_parameter_names(out)
+  out <- merge(params, out, all.x = TRUE)
+  out <- out[order(out$Parameter, params$Parameter), ]
+
+  out <- .remove_backticks_from_parameter_names(out[c("Parameter", "Statistic", "Component")])
   attr(out, "statistic") <- find_statistic(x)
   out
 }